requirements.txt
Django==3.0.8
django-mysql==3.8.1
mysqlclient
djangorestframework
djangorestframework-simplejwt
django-crontab